+++ src/FsCodec.NewtonsoftJson/Codec.fs
@@ -54,25 +54,25 @@ type Codec private () =
 
     static let defaultSettings = lazy Settings.Create()
 
-    /// Generate a <code>IUnionEncoder</code> Codec, using the supplied <c>Newtonsoft.Json<c/> <c>settings</c>.
+    /// Generate an <code>IEventCodec</code> using the supplied <c>Newtonsoft.Json<c/> <c>settings</c>.
     /// Uses <c>up</c> and <c>down</c> functions to facilitate upconversion/downconversion
-    ///   and/or surfacing metadata to the Programming Model by including it in the emitted <c>'Union</c>
+    ///   and/or surfacing metadata to the Programming Model by including it in the emitted <c>'Event</c>
     /// The Event Type Names are inferred based on either explicit <c>DataMember(Name=</c> Attributes, or (if unspecified) the Discriminated Union Case Name
     /// <c>Contract</c> must be tagged with </c>interface TypeShape.UnionContract.IUnionContract</c> to signify this scheme applies.
-    static member Create<'Union,'Contract,'Meta,'Context when 'Contract :> TypeShape.UnionContract.IUnionContract>
+    static member Create<'Event,'Contract,'Meta,'Context when 'Contract :> TypeShape.UnionContract.IUnionContract>
         (   /// Maps from the TypeShape <c>UnionConverter</c> <c>'Contract</c> case the Event has been mapped to (with the raw event data as context)
-            /// to the representation (typically a Discriminated Union) that is to be presented to the programming model.
-            up : FsCodec.ITimelineEvent<byte[]> * 'Contract -> 'Union,
+            /// to the <c>'Event</c> representation (typically a Discriminated Union) that is to be presented to the programming model.
+            up : FsCodec.ITimelineEvent<byte[]> * 'Contract -> 'Event,
             /// Maps a fresh Event resulting from a Decision in the Domain representation type down to the TypeShape <c>UnionConverter</c> <c>'Contract</c>
             /// The function is also expected to derive
             ///   a <c>meta</c> object that will be serialized with the same settings (if it's not <c>None</c>)
             ///   and an Event Creation <c>timestamp</c>.
-            down : 'Context option * 'Union -> 'Contract * 'Meta option * string * string * DateTimeOffset option,
+            down : 'Context option * 'Event -> 'Contract * 'Meta option * string * string * DateTimeOffset option,
             /// Configuration to be used by the underlying <c>Newtonsoft.Json</c> Serializer when encoding/decoding. Defaults to same as <c>Settings.Create()</c>
             [<Optional;DefaultParameterValue(null)>]?settings,
             /// Enables one to fail encoder generation if union contains nullary cases. Defaults to <c>false</c>, i.e. permitting them
             [<Optional;DefaultParameterValue(null)>]?rejectNullaryCases)
-        : FsCodec.IUnionEncoder<'Union,byte[],'Context> =
+        : FsCodec.IEventCodec<'Event,byte[],'Context> =
         let settings = match settings with Some x -> x | None -> defaultSettings.Value
         let bytesEncoder : TypeShape.UnionContract.IEncoder<_> = new Core.BytesEncoder(settings) :> _
         let dataCodec =
@@ -80,66 +80,66 @@ type Codec private () =
                 bytesEncoder,
                 requireRecordFields=true, // See JsonConverterTests - round-tripping UTF-8 correctly with Json.net is painful so for now we lock up the dragons
                 allowNullaryCases=not (defaultArg rejectNullaryCases false))
-        { new FsCodec.IUnionEncoder<'Union,byte[],'Context> with
-            member __.Encode(context,u) =
-                let (c, meta : 'Meta option, correlationId, causationId, timestamp : DateTimeOffset option) = down (context,u)
+        { new FsCodec.IEventCodec<'Event,byte[],'Context> with
+            member __.Encode(context,event) =
+                let (c, meta : 'Meta option, correlationId, causationId, timestamp : DateTimeOffset option) = down (context, event)
                 let enc = dataCodec.Encode c
                 let metaUtf8 = meta |> Option.map bytesEncoder.Encode<'Meta>
-                FsCodec.Core.EventData.Create(enc.CaseName, enc.Payload, defaultArg metaUtf8 null, correlationId, causationId, ?timestamp=timestamp) :> _
+                FsCodec.Core.EventData.Create(enc.CaseName, enc.Payload, defaultArg metaUtf8 null, correlationId, causationId, ?timestamp=timestamp)
             member __.TryDecode encoded =
                 let cOption = dataCodec.TryDecode { CaseName = encoded.EventType; Payload = encoded.Data }
                 match cOption with None -> None | Some contract -> let union = up (encoded,contract) in Some union }
 
-    /// Generate a <code>IUnionEncoder</code> Codec, using the supplied <c>Newtonsoft.Json<c/> <c>settings</c>.
+    /// Generate an <code>IEventCodec</code> using the supplied <c>Newtonsoft.Json<c/> <c>settings</c>.
     /// Uses <c>up</c> and <c>down</c> and <c>mapCausation</c> functions to facilitate upconversion/downconversion and correlation/causationId mapping
-    ///   and/or surfacing metadata to the Programming Model by including it in the emitted <c>'Union</c>
+    ///   and/or surfacing metadata to the Programming Model by including it in the emitted <c>'Event</c>
     /// The Event Type Names are inferred based on either explicit <c>DataMember(Name=</c> Attributes, or (if unspecified) the Discriminated Union Case Name
     /// <c>Contract</c> must be tagged with </c>interface TypeShape.UnionContract.IUnionContract</c> to signify this scheme applies.
-    static member Create<'Union,'Contract,'Meta,'Context when 'Contract :> TypeShape.UnionContract.IUnionContract>
+    static member Create<'Event,'Contract,'Meta,'Context when 'Contract :> TypeShape.UnionContract.IUnionContract>
         (   /// Maps from the TypeShape <c>UnionConverter</c> <c>'Contract</c> case the Event has been mapped to (with the raw event data as context)
             /// to the representation (typically a Discriminated Union) that is to be presented to the programming model.
-            up : FsCodec.ITimelineEvent<byte[]> * 'Contract -> 'Union,
+            up : FsCodec.ITimelineEvent<byte[]> * 'Contract -> 'Event,
             /// Maps a fresh Event resulting from a Decision in the Domain representation type down to the TypeShape <c>UnionConverter</c> <c>'Contract</c>
             /// The function is also expected to derive
             ///   a <c>meta</c> object that will be serialized with the same settings (if it's not <c>None</c>)
             ///   and an Event Creation <c>timestamp</c>.
-            down : 'Union -> 'Contract * 'Meta option * DateTimeOffset option,
+            down : 'Event -> 'Contract * 'Meta option * DateTimeOffset option,
             /// Uses the 'Context passed to the Encode call and the 'Meta emitted by <c>down</c> to a) the final metadata b) the <c>correlationId</c> and c) the correlationId
             mapCausation : 'Context option * 'Meta option -> 'Meta option * string * string,
             /// Configuration to be used by the underlying <c>Newtonsoft.Json</c> Serializer when encoding/decoding. Defaults to same as <c>Settings.Create()</c>
             [<Optional;DefaultParameterValue(null)>]?settings,
             /// Enables one to fail encoder generation if union contains nullary cases. Defaults to <c>false</c>, i.e. permitting them
             [<Optional;DefaultParameterValue(null)>]?rejectNullaryCases)
-        : FsCodec.IUnionEncoder<'Union,byte[],'Context> =
+        : FsCodec.IEventCodec<'Event,byte[],'Context> =
         let down (context,union) =
             let c, m, t = down union
             let m', correlationId, causationId = mapCausation (context,m)
             c, m', correlationId, causationId, t
         Codec.Create(up=up, down=down, ?settings=settings, ?rejectNullaryCases=rejectNullaryCases)
 
-    /// Generate a <code>IUnionEncoder</code> Codec, using the supplied <c>Newtonsoft.Json<c/> <c>settings</c>.
+    /// Generate an <code>IEventCodec</code> using the supplied <c>Newtonsoft.Json<c/> <c>settings</c>.
     /// Uses <c>up</c> and <c>down</c> and <c>mapCausation</c> functions to facilitate upconversion/downconversion and correlation/causationId mapping
-    ///   and/or surfacing metadata to the Programming Model by including it in the emitted <c>'Union</c>
+    ///   and/or surfacing metadata to the Programming Model by including it in the emitted <c>'Event</c>
     /// The Event Type Names are inferred based on either explicit <c>DataMember(Name=</c> Attributes, or (if unspecified) the Discriminated Union Case Name
     /// <c>Contract</c> must be tagged with </c>interface TypeShape.UnionContract.IUnionContract</c> to signify this scheme applies.
-    static member Create<'Union,'Contract,'Meta when 'Contract :> TypeShape.UnionContract.IUnionContract>
+    static member Create<'Event,'Contract,'Meta when 'Contract :> TypeShape.UnionContract.IUnionContract>
         (   /// Maps from the TypeShape <c>UnionConverter</c> <c>'Contract</c> case the Event has been mapped to (with the raw event data as context)
             /// to the representation (typically a Discriminated Union) that is to be presented to the programming model.
-            up : FsCodec.ITimelineEvent<byte[]> * 'Contract -> 'Union,
-            /// Maps a fresh Event resulting from a Decision in the Domain representation type down to the TypeShape <c>UnionConverter</c> <c>'Contract</c>
+            up : FsCodec.ITimelineEvent<byte[]> * 'Contract -> 'Event,
+            /// Maps a fresh <c>'Event</c> resulting from a Decision in the Domain representation type down to the TypeShape <c>UnionConverter</c> <c>'Contract</c>
             /// The function is also expected to derive
             ///   a <c>meta</c> object that will be serialized with the same settings (if it's not <c>None</c>)
             ///   and an Event Creation <c>timestamp</c>.
-            down : 'Union -> 'Contract * 'Meta option * DateTimeOffset option,
+            down : 'Event -> 'Contract * 'Meta option * DateTimeOffset option,
             /// Configuration to be used by the underlying <c>Newtonsoft.Json</c> Serializer when encoding/decoding. Defaults to same as <c>Settings.Create()</c>
             [<Optional;DefaultParameterValue(null)>]?settings,
             /// Enables one to fail encoder generation if union contains nullary cases. Defaults to <c>false</c>, i.e. permitting them
             [<Optional;DefaultParameterValue(null)>]?rejectNullaryCases)
-        : FsCodec.IUnionEncoder<'Union,byte[],obj> =
+        : FsCodec.IEventCodec<'Event,byte[],obj> =
         let mapCausation (_context : obj, m : ' Meta option) = m,null,null
         Codec.Create(up=up, down=down, mapCausation=mapCausation, ?settings=settings, ?rejectNullaryCases=rejectNullaryCases)
 
-    /// Generate a <code>IUnionEncoder</code> Codec using the supplied <c>Newtonsoft.Json</c> <c>settings</c>.
+    /// Generate an <code>IEventCodec</code> using the supplied <c>Newtonsoft.Json</c> <c>settings</c>.
     /// The Event Type Names are inferred based on either explicit <c>DataMember(Name=</c> Attributes, or (if unspecified) the Discriminated Union Case Name
     /// <c>'Union</c> must be tagged with <c>interface TypeShape.UnionContract.IUnionContract</c> to signify this scheme applies.
     static member Create<'Union when 'Union :> TypeShape.UnionContract.IUnionContract>
@@ -147,7 +147,7 @@ type Codec private () =
             [<Optional;DefaultParameterValue(null)>]?settings,
             /// Enables one to fail encoder generation if union contains nullary cases. Defaults to <c>false</c>, i.e. permitting them
             [<Optional;DefaultParameterValue(null)>]?rejectNullaryCases)
-        : FsCodec.IUnionEncoder<'Union, byte[], obj> =
+        : FsCodec.IEventCodec<'Union, byte[], obj> =
         let up : FsCodec.ITimelineEvent<_> * 'Union -> 'Union = snd
-        let down (u : 'Union) = u, None, None
+        let down (event : 'Union) = event, None, None
         Codec.Create(up=up, down=down, ?settings=settings, ?rejectNullaryCases=rejectNullaryCases)