+++ tests/FsCodec.NewtonsoftJson.Tests/VerbatimUtf8ConverterTests.fs
@@ -55,37 +55,37 @@ let defaultSettings = Settings.CreateDefault()
 #nowarn "1182" // From hereon in, we may have some 'unused' privates (the tests)
 
 type VerbatimUtf8Tests() =
-    let unionEncoder = Codec.Create()
+    let eventCodec = Codec.Create()
 
     [<Fact>]
     let ``encodes correctly`` () =
-        let encoded = unionEncoder.Encode(None, A { embed = "\"" })
+        let encoded = eventCodec.Encode(None, A { embed = "\"" })
         let e : Batch =
             {   p = "streamName"; id = string 0; i = -1L; n = -1L; _etag = null
                 e = [| { t = DateTimeOffset.MinValue; c = encoded.EventType; d = encoded.Data; m = null } |] }
         let res = JsonConvert.SerializeObject(e)
         test <@ res.Contains """"d":{"embed":"\""}""" @>
 
-    let uEncoder = Codec.Create<U>(defaultSettings)
+    let defaultEventCodec = Codec.Create<U>(defaultSettings)
 
     let [<Property(MaxTest=100)>] ``round-trips diverse bodies correctly`` (x: U) =
-        let encoded = uEncoder.Encode(None,x)
+        let encoded = defaultEventCodec.Encode(None,x)
         let e : Batch =
             {   p = "streamName"; id = string 0; i = -1L; n = -1L; _etag = null
                 e = [| { t = DateTimeOffset.MinValue; c = encoded.EventType; d = encoded.Data; m = null } |] }
         let ser = JsonConvert.SerializeObject(e, defaultSettings)
         let des = JsonConvert.DeserializeObject<Batch>(ser, defaultSettings)
         let loaded = FsCodec.Core.TimelineEvent.Create(-1L, des.e.[0].c, des.e.[0].d)
-        let decoded = uEncoder.TryDecode loaded |> Option.get
+        let decoded = defaultEventCodec.TryDecode loaded |> Option.get
         x =! decoded
 
     // NB while this aspect works, we don't support it as it gets messy when you then use the VerbatimUtf8Converter
     // https://github.com/JamesNK/Newtonsoft.Json/issues/862 // doesnt apply to this case
     let [<Fact>] ``Codec does not fall prey to Date-strings being mutilated`` () =
         let x = ES { embed = "2016-03-31T07:02:00+07:00" }
-        let encoded = uEncoder.Encode(None,x)
+        let encoded = defaultEventCodec.Encode(None,x)
         let adapted = FsCodec.Core.TimelineEvent.Create(-1L, encoded.EventType, encoded.Data, encoded.Meta, timestamp = encoded.Timestamp)
-        let decoded = uEncoder.TryDecode adapted |> Option.get
+        let decoded = defaultEventCodec.TryDecode adapted |> Option.get
         test <@ x = decoded @>
 
     //// NB while this aspect works, we don't support it as it gets messy when you then use the VerbatimUtf8Converter